Moved toggle dark mode to the navbar folder
